Update HQTB #17 #7
- combined modules of refinement into one
- fixed some bugs
- removed a couple of TODOs and the like
