--- src/specimen/hqtb/core/refinement/__init__.py
@@ -1,3 +1,0 @@
-__all__ = ['extension','cleanup','annotation','smoothing']
-
-from . import extension, cleanup, annotation, smoothing
